modularization of js code
