+++ index.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Poch'Lib</title>
     <link rel="stylesheet" href="style.css">
-
 </head>
 <body>
     <div id="myBooks">
@@ -27,6 +26,6 @@
             <ul id="books"></ul>
         </div>
     </div>
-    <script src="app.js"></script>
+    <script type="module" src="app.js"></script>
 </body>
 </html>
